+++ changelog.md
@@ -4,6 +4,7 @@
 
 - Remove shim.js of lib-wrapper, just use the module
 - Bug fix for [BUG: If out of reach, door still opens, revealing vision](https://github.com/p4535992/foundryvtt-arms-reach/issues/1)
+- Add new settings for avoid deselects the controlled token when open/close the door
 
 # 1.0.3